requirements.txt
streamlit == 1.36.0
paho-mqtt == 2.1.0